+++ src/main/java/org/braekpo1nt/mctmanager/games/gamemanager/event/EventData.java
@@ -1,8 +1,7 @@
-package org.braekpo1nt.mctmanager.games.gamemanager.states.event;
+package org.braekpo1nt.mctmanager.games.gamemanager.event;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.braekpo1nt.mctmanager.games.gamemanager.event.ScoreKeeper;
 import org.braekpo1nt.mctmanager.games.gamemanager.event.config.EventConfig;
 import org.braekpo1nt.mctmanager.games.game.enums.GameType;
 import org.braekpo1nt.mctmanager.participant.Participant;
