+++ src/main/java/org/braekpo1nt/mctmanager/games/gamemanager/states/event/PlayingFinalGameState.java
@@ -3,6 +3,7 @@ package org.braekpo1nt.mctmanager.games.gamemanager.states.event;
 import org.braekpo1nt.mctmanager.games.GameManager;
 import org.braekpo1nt.mctmanager.games.game.enums.GameType;
 import org.braekpo1nt.mctmanager.games.gamemanager.MCTTeam;
+import org.braekpo1nt.mctmanager.games.gamemanager.event.EventData;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.ContextReference;
 import org.jetbrains.annotations.NotNull;
