+++ src/main/java/org/braekpo1nt/mctmanager/games/gamemanager/states/event/PlayingGameState.java
@@ -9,6 +9,7 @@ import org.braekpo1nt.mctmanager.commands.manager.commandresult.SuccessCommandRe
 import org.braekpo1nt.mctmanager.games.GameManager;
 import org.braekpo1nt.mctmanager.games.game.enums.GameType;
 import org.braekpo1nt.mctmanager.games.gamemanager.MCTParticipant;
+import org.braekpo1nt.mctmanager.games.gamemanager.event.EventData;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.ContextReference;
 import org.jetbrains.annotations.NotNull;
