+++ src/main/java/org/braekpo1nt/mctmanager/games/gamemanager/states/event/delay/DelayState.java
@@ -4,7 +4,7 @@ import org.braekpo1nt.mctmanager.commands.manager.commandresult.CommandResult;
 import org.braekpo1nt.mctmanager.games.GameManager;
 import org.braekpo1nt.mctmanager.games.game.enums.GameType;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.ContextReference;
-import org.braekpo1nt.mctmanager.games.gamemanager.states.event.EventData;
+import org.braekpo1nt.mctmanager.games.gamemanager.event.EventData;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.event.EventState;
 import org.jetbrains.annotations.NotNull;
