+++ src/main/java/org/braekpo1nt/mctmanager/games/gamemanager/states/event/delay/StartingGameDelayState.java
@@ -4,7 +4,7 @@ import net.kyori.adventure.text.Component;
 import org.braekpo1nt.mctmanager.games.GameManager;
 import org.braekpo1nt.mctmanager.games.game.enums.GameType;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.ContextReference;
-import org.braekpo1nt.mctmanager.games.gamemanager.states.event.EventData;
+import org.braekpo1nt.mctmanager.games.gamemanager.event.EventData;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.event.PlayingGameState;
 import org.braekpo1nt.mctmanager.ui.timer.Timer;
 import org.jetbrains.annotations.NotNull;
